Refactor fork config
