+++ radixdlt-core/radixdlt/src/main/java/com/radixdlt/statecomputer/RadixEngineModule.java
@@ -25,18 +25,23 @@ import com.google.inject.multibindings.Multibinder;
 import com.radixdlt.atom.ActionConstructors;
 import com.radixdlt.atommodel.system.state.SystemParticle;
 import com.radixdlt.atommodel.system.state.ValidatorEpochData;
+import com.radixdlt.consensus.LedgerProof;
+import com.radixdlt.consensus.bft.View;
 import com.radixdlt.constraintmachine.ConstraintMachine;
 import com.radixdlt.engine.PostParsedChecker;
 import com.radixdlt.engine.BatchVerifier;
 import com.radixdlt.engine.RadixEngine;
 import com.radixdlt.engine.StateReducer;
 import com.radixdlt.engine.SubstateCacheRegister;
+import com.radixdlt.statecomputer.forks.ForkConfig;
 import com.radixdlt.store.EngineStore;
+import com.radixdlt.sync.CommittedReader;
 import com.radixdlt.utils.Pair;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
 import java.util.Set;
+import java.util.TreeMap;
 
 /**
  * Module which manages execution of commands
@@ -50,6 +55,53 @@ public class RadixEngineModule extends AbstractModule {
 		Multibinder.newSetBinder(binder(), new TypeLiteral<Pair<String, StateReducer<?>>>() { });
 		Multibinder.newSetBinder(binder(), PostParsedChecker.class);
 		Multibinder.newSetBinder(binder(), new TypeLiteral<SubstateCacheRegister<?>>() { });
+	}
+
+
+	@Provides
+	@Singleton
+	@EpochCeilingView
+	private View epochCeilingHighView(
+		CommittedReader committedReader, // TODO: This is a hack, remove
+		TreeMap<Long, ForkConfig> epochToForkConfig
+	) {
+		var lastProof = committedReader.getLastProof().orElse(LedgerProof.mock());
+		var epoch = lastProof.isEndOfEpoch() ? lastProof.getEpoch() + 1 : lastProof.getEpoch();
+		return epochToForkConfig.floorEntry(epoch).getValue().getEpochCeilingView();
+	}
+
+	@Provides
+	@Singleton
+	private ConstraintMachine buildConstraintMachine(
+		CommittedReader committedReader, // TODO: This is a hack, remove
+		TreeMap<Long, ForkConfig> epochToForkConfig
+	) {
+		var lastProof = committedReader.getLastProof().orElse(LedgerProof.mock());
+		var epoch = lastProof.isEndOfEpoch() ? lastProof.getEpoch() + 1 : lastProof.getEpoch();
+		return epochToForkConfig.floorEntry(epoch).getValue().getConstraintMachine();
+	}
+
+	@Provides
+	@Singleton
+	private ActionConstructors actionConstructors(
+		CommittedReader committedReader, // TODO: This is a hack, remove
+		TreeMap<Long, ForkConfig> epochToForkConfig
+	) {
+		var lastProof = committedReader.getLastProof().orElse(LedgerProof.mock());
+		var epoch = lastProof.isEndOfEpoch() ? lastProof.getEpoch() + 1 : lastProof.getEpoch();
+		return epochToForkConfig.floorEntry(epoch).getValue().getActionConstructors();
+	}
+
+
+	@Provides
+	@Singleton
+	private BatchVerifier<LedgerAndBFTProof> batchVerifier(
+		CommittedReader committedReader, // TODO: This is a hack, remove
+		TreeMap<Long, ForkConfig> epochToForkConfig
+	) {
+		var lastProof = committedReader.getLastProof().orElse(LedgerProof.mock());
+		var epoch = lastProof.isEndOfEpoch() ? lastProof.getEpoch() + 1 : lastProof.getEpoch();
+		return epochToForkConfig.floorEntry(epoch).getValue().getBatchVerifier();
 	}
 
 	@Provides
